requirements.txt
pyautogui
tkinter
Pillow
keyboard